[TAN-1402] Rename more ClaveUnica steps
