+++ front/app/containers/Authentication/Modal.tsx
@@ -22,11 +22,11 @@ import messages from './messages';
 import AuthProviders from './steps/AuthProviders';
 import BuiltInFields from './steps/BuiltInFields';
 import ChangeEmail from './steps/ChangeEmail';
-import EmaillessSsoEmail from './steps/ClaveUnicaEmail';
 import CustomFields from './steps/CustomFields';
 import EmailAndPassword from './steps/EmailAndPassword';
 import EmailAndPasswordSignUp from './steps/EmailAndPasswordSignUp';
 import EmailConfirmation from './steps/EmailConfirmation';
+import EmaillessSsoEmail from './steps/EmaillessSsoEmail';
 import Invitation from './steps/Invitation';
 import LightFlowStart from './steps/LightFlowStart';
 import Onboarding from './steps/Onboarding';
